requirements.txt
numpy==1.21
aiohttp==3.8.1
async-timeout==4.0.2
chardet==3.0.4
multidict==4.7.5
python-engineio
python_socketio
six==1.14.0
yarl==1.4.2
click
pyyaml
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.2  
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.2 

matplotlib
sklearn
fdasrsf
seaborn 
mne